+++ extensions/infisical/src/secrets.tsx
@@ -239,7 +239,7 @@ export default function Secrets({
                   title={revealValues ? "Hide Values" : "Reveal Values"}
                   onAction={() => setRevealValues((reveal) => !reveal)}
                 />
-                <Action.CopyToClipboard title="Copy Secret" content={secret.secretValue} />
+                <Action.CopyToClipboard title="Copy Secret" content={secret.secretValue} concealed />
                 <Action.Push
                   icon={Icon.Pencil}
                   title="Edit Secret"
@@ -266,6 +266,7 @@ export default function Secrets({
                   // eslint-disable-next-line @raycast/prefer-title-case
                   title="Copy All as .env"
                   content={combineSecretsAsEnv(secrets)}
+                  concealed
                   shortcut={Keyboard.Shortcut.Common.Copy}
                 />
                 <Action